+++ search-engine-spring-boot/src/main/java/com/searchengine/dao/RecordSegDao.java
@@ -12,9 +12,9 @@ public interface RecordSegDao {
     //查看所有关系
     List<RecordSeg> selectAllRecordSeg();
     //根据recordid和segid查询单个关系
-    RecordSeg selectOneRecordSeg(@Param("dataId") Long recordId, @Param("segId") Long segId);
+    RecordSeg selectOneRecordSeg(@Param("dataId") Integer recordId, @Param("segId") Integer segId);
     //根据segId查找对应的recordId
-    List<RecordSeg> selectRecordBySeg(@Param("segId") Long segId);
+    List<RecordSeg> selectRecordBySeg(@Param("segId") Integer segId);
     //存储关系表
     int insertRecordSeg(RecordSeg recordSeg);
     //更新关系表
